Create README.md
Hi, I’m James and living in Florida. I’m a well-known writer and known to be the author of most of the articles published on various fishing reel sites. I am an avid fisherman and expert when it comes to fishing reels. I’m also a part-time lecturer in the university and conduct classes on fishing reels and fishing rods.